+++ pkg/terraform/state/deploy.go
@@ -1,4 +1,4 @@
-package execution
+package state
 
 import (
 	"bytes"
